Update real-time-data-ingestion.md
UI did change: We added content into that page for all API usage. As part of that we moved the real-time API page into a side panel. I tried to explain how this work (it's clunky to navigate to it). I would assume we should link to the page explaining the new content of the main page, likely this should be done in this UA page: https://docs.microsoft.com/en-us/dynamics365/customer-insights/audience-insights/apis - currently there's no mention of it.