Update export-facebook.md
Change this export type as a sample for how others will be structured. Idea is: First the basics like prereq and limitations, then connection, then export. IMO export and connection could swap, but why not stick with what's necessary first - even if we assume a whole set of users will never setup the connection themselves?