Adding recommendations to D365 MKT export
We have recurring Cx questions and problems bc of the MKT export. Most of the issues stem from MKT's API not being performant and robust enough for this use case. I added a heading "recommendations" with two bullet points as content. 
I also changed the 2nd paragraph on the page to emphasis that Cx should consider using MKT real-time instead of outbound.